# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Virtual Care API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same host"
    },
    {
      "url": "https://agent.fly.dev",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    }
  ],
...
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/reports/summary": {
      "get": {
        "operationId": "getSummary",
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "groupBy",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["client", "project"]
            }
          },
          {
            "in": "query",
            "name": "clientId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "tags": ["reports"],
        "description": "Get summary report of hours and amounts by client or project for a date range.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "groupBy": {
                          "enum": ["client"],
                          "type": "string"
                        },
                        "dateRange": {
                          "type": "object",
                          "properties": {
                            "startDate": {
                              "type": "string"
                            },
                            "endDate": {
                              "type": "string"
                            }
                          },
                          "required": ["startDate", "endDate"],
                          "additionalProperties": false
                        },
                        "summary": {
                          "type": "array",
                          "items": {}
                        }
                      },
                      "required": ["groupBy", "dateRange", "summary"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "groupBy": {
                          "enum": ["project"],
                          "type": "string"
                        },
                        "dateRange": {
                          "type": "object",
                          "properties": {
                            "startDate": {
                              "type": "string"
                            },
                            "endDate": {
                              "type": "string"
                            }
                          },
                          "required": ["startDate", "endDate"],
                          "additionalProperties": false
                        },
                        "summary": {
                          "type": "array",
                          "items": {}
                        }
                      },
                      "required": ["groupBy", "dateRange", "summary"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "error": {
                            "enum": ["End date must be after start date"],
                            "type": "string"
                          }
                        },
                        "required": ["error"],
                        "additionalProperties": false
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/detailed": {
      "get": {
        "operationId": "getDetailedReport",
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "clientId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "tags": ["reports"],
        "description": "Get detailed time entries for a specific date range with totals.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "durationHours": {
                            "type": "number"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "project": {
                            "allOf": [
                              {
                                "type": "object",
                                "properties": {
                                  "client": {
                                    "anyOf": [
                                      {
                                        "type": "null"
                                      },
                                      {
                                        "type": "object",
                                        "properties": {
                                          "id": {
                                            "type": "string"
                                          },
                                          "name": {
                                            "type": "string"
                                          },
                                          "email": {
                                            "anyOf": [
                                              {
                                                "type": "null"
                                              },
                                              {
                                                "type": "string"
                                              }
                                            ]
                                          },
                                          "createdAt": {
                                            "type": "string"
                                          },
                                          "updatedAt": {
                                            "type": "string"
                                          },
                                          "userId": {
                                            "type": "string"
                                          }
                                        },
                                        "required": [
                                          "id",
                                          "name",
                                          "email",
                                          "createdAt",
                                          "updatedAt",
                                          "userId"
                                        ],
                                        "additionalProperties": false
                                      }
                                    ]
                                  }
                                },
                                "required": ["client"],
                                "additionalProperties": false
                              },
                              {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "name": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string"
                                  },
                                  "updatedAt": {
                                    "type": "string"
                                  },
                                  "userId": {
                                    "type": "string"
                                  },
                                  "description": {
                                    "anyOf": [
                                      {
                                        "type": "null"
                                      },
                                      {
                                        "type": "string"
                                      }
                                    ]
                                  },
                                  "hourlyRate": {
                                    "anyOf": [
                                      {
                                        "type": "null"
                                      },
                                      {
                                        "type": "string"
                                      }
                                    ]
                                  },
                                  "clientId": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "id",
                                  "name",
                                  "createdAt",
                                  "updatedAt",
                                  "userId",
                                  "description",
                                  "hourlyRate",
                                  "clientId"
                                ],
                                "additionalProperties": false
                              }
                            ]
                          },
                          "id": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "startedAt": {
                            "type": "string"
                          },
                          "endedAt": {
                            "type": "string"
                          },
                          "note": {
                            "anyOf": [
                              {
                                "type": "null"
                              },
                              {
                                "type": "string"
                              }
                            ]
                          },
                          "isLocked": {
                            "type": "boolean"
                          },
                          "projectId": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "durationHours",
                          "amount",
                          "project",
                          "id",
                          "createdAt",
                          "updatedAt",
                          "userId",
                          "startedAt",
                          "endedAt",
                          "note",
                          "isLocked",
                          "projectId"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "pageSize": {
                          "type": "number"
                        },
                        "totalCount": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "page",
                        "pageSize",
                        "totalCount",
                        "totalPages"
                      ],
                      "additionalProperties": false
                    },
                    "totals": {
                      "type": "object",
                      "properties": {
                        "totalHours": {
                          "type": "number"
                        },
                        "totalAmount": {
                          "type": "number"
                        }
                      },
                      "required": ["totalHours", "totalAmount"],
                      "additionalProperties": false
                    },
                    "dateRange": {
                      "type": "object",
                      "properties": {
                        "startDate": {
                          "type": "string"
                        },
                        "endDate": {
                          "type": "string"
                        }
                      },
                      "required": ["startDate", "endDate"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["data", "pagination", "totals", "dateRange"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UnauthorizedErr": {
        "type": "object",
        "required": ["type", "title"],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "enum": ["Unauthorized"]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}